# crates/quest/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CMU Orientation Quest",
    "description": "Backend for the Orientation Quest app.",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/logout": {
      "post": {
        "tags": ["auth"],
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "502": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          }
        }
      }
    },
    "/auth/status": {
      "get": {
        "tags": ["auth"],
        "operationId": "status",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserView"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "502": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthErrBody": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "LogoutResponse": {
        "type": "object",
        "properties": {
          "end_session_url": {
            "type": ["string", "null"]
          }
        }
      },
...
      "UserView": {
        "type": "object",
        "required": ["name", "andrew_id", "groups", "admin", "staff"],
        "properties": {
          "admin": {
            "type": "boolean"
          },
          "andrew_id": {
            "type": "string"
          },
          "email": {
            "type": ["string", "null"]
          },
          "groups": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "staff": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "device_proof": {
        "type": "apiKey",
        "in": "header",
        "name": "x-device-proof",
        "description": "Signature over the request method and URL by an enrolled device key"
      },
      "session": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session id issued by /auth/login"
      }
    }
  },
  "security": [
    {
      "device_proof": [],
      "session": []
    }
  ],
...
}